+++ src/ladb_opencutlist/ruby/controller/cutlist_controller.rb
@@ -1010,7 +1010,7 @@ module Ladb::OpenCutList
         response[:groups].push(group)
 
         # Folding
-        if part_folding and group_def.material_type > MaterialAttributes::TYPE_UNKNOW   # Only parts with typed material can be grouped
+        if part_folding # and group_def.material_type > MaterialAttributes::TYPE_UNKNOW   # Only parts with typed material can be grouped
           part_defs = []
           group_def.part_defs.values.sort_by { |v| [ v.size.thickness, v.size.length, v.size.width, v.labels, v.final_area ] }.each { |part_def|
             if !(folder_part_def = part_defs.last).nil? &&
@@ -1018,7 +1018,8 @@ module Ladb::OpenCutList
                 folder_part_def.cutting_size == part_def.cutting_size &&
                 (folder_part_def.labels == part_def.labels || hide_labels) &&
                 ((folder_part_def.final_area - part_def.final_area).abs < 0.001 or hide_final_areas) &&      # final_area workaround for rounding error
-                folder_part_def.edge_material_names == part_def.edge_material_names
+                folder_part_def.edge_material_names == part_def.edge_material_names &&
+                ((folder_part_def.definition_id == part_def.definition_id && group_def.material_type == MaterialAttributes::TYPE_UNKNOW) || group_def.material_type > MaterialAttributes::TYPE_UNKNOW) # Part with untyped materiel are folded only if they have the same definition
               if folder_part_def.children.empty?
                 first_child_part_def = part_defs.pop